moved the user route to authentication_controller to avoid having so many user routes and this controller was all about login
